refactor(finance): extract imports service pipeline stages (#1718)
Extract 4 modules from service.ts (1411 → 634 LOC) into lib/:
- deduplication.ts: findExistingChecksums()
- tag-management.ts: parseCorrectionTags(), loadKnownTags(), buildSuggestedTags()
- correction-application.ts: applyLearnedCorrection(), reevaluateImportSessionResult(), reevaluateImportSessionWithRules()
- transaction-persistence.ts: insertTransaction(), commitImport(), validateCommitPayload(), resolveChangeSetTempIds(), reclassifyExistingTransactions()

service.ts retains pipeline orchestrators only. All existing tests pass unchanged.
